Rename file to be consistent with class
