+++ example/lib/main.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 
 import 'constants.dart';
 import 'op_state.dart';
-import 'push_notifications/push_notification_event_handlers.dart';
+import 'push_notifications/push_notification_handlers.dart';
 import 'push_notifications/push_notification_service.dart';
 import 'ui/push_notifications/push_notifications_sliver.dart';
